Import Grunt automation toolkit, and contributed linters to help catch errors early.
This will also enforce a coding style between both the coffeescript and
javascript libraries in the repository. Any contributions should conform
to the standards.

Warehousing this in the repo vs the Dockerfile repository ensures that
we get consistent results across the board from Developers and CI. If
you find a reson for the Linters to be turned off - I'm all ears to hear
it, but I doubt you'll win me over :)