Updated all *.coffee scripts
coffeelint is aggressive in style, line length at 80 characters, and
enforcing consistent indentation/semicolon use.

Most of the imported scripts were riddled with whitespace/indentation
errors. This is a mass cleanup of those errors - be sure you pay special
attention to the refactorings of janky.coffee and other method signature
breaks - I'm not sure if this will completely break the scripts or not.

This was my best attempt at following the coding styles of coffeelint

Also removed:

- example.coffee
- commandlinefu (stupid indentation/format - too intelligent for its own
  good.